Rename wrapper to row_converter
