+++ datafusion/core/src/physical_plan/mod.rs
@@ -402,7 +402,7 @@ pub mod union;
 pub mod unnest;
 pub mod values;
 pub mod windows;
-mod wrapper;
+mod row_converter;
 
 use crate::physical_plan::repartition::RepartitionExec;
 use crate::physical_plan::sorts::sort_preserving_merge::SortPreservingMergeExec;
