+++ datafusion/core/src/physical_plan/sorts/stream.rs
@@ -16,7 +16,7 @@
 // under the License.
 
 use crate::physical_plan::sorts::cursor::{FieldArray, FieldCursor, RowCursor};
-use crate::physical_plan::wrapper::CardinalityAwareRowConverter;
+use crate::physical_plan::row_converter::CardinalityAwareRowConverter;
 use crate::physical_plan::SendableRecordBatchStream;
 use crate::physical_plan::{PhysicalExpr, PhysicalSortExpr};
 use arrow::array::Array;
